+++ frontend/src/utils/consts.ts
@@ -124,7 +124,8 @@ export const TEXTS_HEBREW = {
     DAILYSUMMARYFOROBSERVER: "סיכום טיסות למפקח",
     SUMMARYFORTOWPILOT: "סיכום טיסות לטייס גורר",
     SUMMARYFORCFI: "סיכום טיסות למדריך",
-    PENDING: "ממתין לשליחה",
+    HANDLED: "טופל",
+    PENDING: "ממתין לטיפול",
     BEING_HANDLED: "בטיפול",
     SENT: "נשלח",
     FAILED: "נכשל",
@@ -140,7 +141,11 @@ export const TEXTS_HEBREW = {
     CONFIRM_ACTION_REOPEN: 'האם אתה בטוח שברצונך לפתוח מחדש את הפעולה?',
     REOPEN_ACTION: 'פתיחה מחדש של הפעולה',
     REOPEN_ACTION_EVENT_CONFIRMATION: 'האם ברצונך ליצור אירוע פתיחה מחדש?',
+    EXPORT_ACTION_DATA: 'ייצוא נתוני הפעולה',
+    ACTION_DATA_EXPORTED_AT: 'נתוני הפעולה יוצאו בתאריך',
+    ACTION_DATA_NOT_YET_EXPORTED: 'נתוני הפעולה טרם יוצאו',
     ACTION_REOPENED: 'הפעולה נפתחה מחדש',
+    CONFIRM_ACTION_DATA_EXPORT: 'האם אתה בטוח שברצונך לבקש את ייצוא נתוני הפעולה?',
     SETTLE_PAYMENT: 'הסדר תשלום',
     PAYMENT_BY_CLUB_MEMBER: 'חבר מועדון משלם על הטיסה',
     FLIGHTS_WITH_UNSETTLED_PAYMENTS_TITLE: 'קיימות טיסות עם תשלומים שלא הוסדרו',
@@ -152,6 +157,10 @@ export const TEXTS_HEBREW = {
     NO_FLIGHTS_IN_THIS_STATE: 'אין טיסות במצב זה',
     RESEND_NOTIFICATION: 'שלח מחדש',
     RESEND_NOTIFICATION_CONFIRMATION: 'האם אתה בטוח שברצונך לשלוח מחדש את ההודעה?',
+    DATA_EXPORT_REQUEST_EVENTS: 'בקשות ייצוא נתונים',
+    NO_DATA_EXPORT_REQUEST_EVENTS_FOR_ACTION: 'אין בקשות ייצוא נתונים עבור הפעולה',
+    ACTION_DATE: 'תאריך הפעולה',
+    ACTION_EXPORTED_AT: 'נתוני הפעולה יוצאו בתאריך',
 }